requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Actor System (Akka for Python)
pykka==4.0.1

# Event Streaming (Redpanda/Kafka)
# Using kafka-python-ng (maintained fork for Python 3.12+)
kafka-python-ng==2.2.2

# Authentication (Auth0)
authlib==1.3.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# ML Observability (Comet)
comet-ml==3.35.3

# LLM & Agent Framework
langchain==0.1.0
langgraph==0.0.20
openai==1.6.1
anthropic==0.8.1

# Utilities
python-dotenv==1.0.0
httpx==0.25.2

# Development & Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
black==23.12.1
ruff==0.1.8
mypy==1.7.1

# Type stubs
types-requests==2.31.0.10
